Revert "LPS-108310 portal-search-elasticsearch7-impl: extract logic into IndexSettingsContributorHelper, prepare for next"
This reverts commit 9ec6580030a532edca7211d0107957a82424699c.
