--- modules/apps/portal-search-elasticsearch7/portal-search-elasticsearch7-impl/src/main/java/com/liferay/portal/search/elasticsearch7/internal/settings/IndexSettingsContributorHelper.java
@@ -1,87 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.portal.search.elasticsearch7.internal.settings;
-
-import com.liferay.portal.search.elasticsearch7.internal.index.LiferayDocumentTypeFactory;
-import com.liferay.portal.search.elasticsearch7.settings.IndexSettingsContributor;
-import com.liferay.portal.search.elasticsearch7.settings.IndexSettingsHelper;
-
-import java.util.Set;
-import java.util.concurrent.ConcurrentSkipListSet;
-
-import org.elasticsearch.common.settings.Settings;
-
-import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
-import org.osgi.service.component.annotations.ReferenceCardinality;
-import org.osgi.service.component.annotations.ReferencePolicy;
-import org.osgi.service.component.annotations.ReferencePolicyOption;
-
-/**
- * @author Tina Tian
- */
-@Component(service = IndexSettingsContributorHelper.class)
-public class IndexSettingsContributorHelper {
-
-	public void loadIndexSettingsContributors(Settings.Builder builder) {
-		IndexSettingsHelper indexSettingsHelper = new IndexSettingsHelper() {
-
-			@Override
-			public void put(String setting, String value) {
-				builder.put(setting, value);
-			}
-
-		};
-
-		for (IndexSettingsContributor indexSettingsContributor :
-				_indexSettingsContributors) {
-
-			indexSettingsContributor.populate(indexSettingsHelper);
-		}
-	}
-
-	public void loadTypeMappingsContributors(
-		String indexName,
-		LiferayDocumentTypeFactory liferayDocumentTypeFactory) {
-
-		for (IndexSettingsContributor indexSettingsContributor :
-				_indexSettingsContributors) {
-
-			indexSettingsContributor.contribute(
-				indexName, liferayDocumentTypeFactory);
-		}
-	}
-
-	@Reference(
-		cardinality = ReferenceCardinality.MULTIPLE,
-		policy = ReferencePolicy.DYNAMIC,
-		policyOption = ReferencePolicyOption.GREEDY
-	)
-	protected void addIndexSettingsContributor(
-		IndexSettingsContributor indexSettingsContributor) {
-
-		_indexSettingsContributors.add(indexSettingsContributor);
-	}
-
-	protected void removeIndexSettingsContributor(
-		IndexSettingsContributor indexSettingsContributor) {
-
-		_indexSettingsContributors.remove(indexSettingsContributor);
-	}
-
-	private final Set<IndexSettingsContributor> _indexSettingsContributors =
-		new ConcurrentSkipListSet<>();
-
-}
